Authorization now filters users without auth

+++ goserver/src/bloomgenetics.tech/bloom/auth/auth.go
@@ -27,7 +27,7 @@ func GetLogin(auth string) entity.UserLogin {
 	return result
 }
 
-func CheckAuth(uid string, pid int, pagename string, method string) (bool, error) {
+func CheckAuth(uid string, pid int64, pagename string, method string) (bool, error) {
 	const qBase = "select page from roles r join role_perm rp using(role_id) join perm p on p.id = rp.perm_id where r.user_id = $1 and r.project_id = $2 and p.page = $3 and p.action = $4"
 	rows, err := util.Database.Query(qBase, uid, pid, pagename, method)
 	if err != nil {
